* master:
  Add ability to redirect stderr from git hooks
  Add possibility to get pure stderr output from AbortedByHookException
  Fix API check errors introduced in ReceivePack in 945d43d5
  Cleanup CommitAndLogCommandTest
  CLI: Add support for excluding paths from Git log command
  ReceivePack: Moves connectivity check to separate class
  Upgrade error_prone_core to 2.3.4
  Suppress API error for RefDatabase.hasFastTipsWithSha1()
  Remove API filters which became obsolete after bumping version to 5.7.0
  ReftableTest: Clean up boxing warnings on usage of String.format
  Add missing @since tag for ReceivePack#receiveWithExceptionPropagation
  MergedReftable: Remove unnecessary semicolon
  Prepare 5.7.0-SNAPSHOT builds
  Introduce RefDatabase#hasFastTipsWithSha1
  Update Orbit to R20191126223242 for 2019-12
  Prepare 5.6.0-SNAPSHOT builds
  JGit v5.6.0.201912041214-rc1
  server: Add a custom error handler
  transport: Add methods that do not handle exceptions
  transport: Let users specify an error handler for unpacking a pack file
  ReceivePack: Use try-with-resources for PostReceiveHook
  transport: Consolidate status reporting code
  transport: Add ReceiveCommandErrorHandler
  transport: Move exception handler to the caller
  transport: Move handling code to the caller side
  UploadPackServlet: Use uploadWithExceptionPropagation
  GitSmartHttpTools: Do not use sideband when sending an error
  Prepare 5.6.0-SNAPSHOT builds
  JGit v5.6.0.201911271000-m3
  ReceivePack: Open visibility for some methods
  JGit pgm: Format blame output to match canonical git
  Make blame work correctly on merge conflicts
  Bazel: Use java_plugin and java_binary from @rules_java in jmh.bzl
  Bazel: Add missing newlines at end of BUILD files
  Upgrade maven-enforcer-plugin to 3.0.0-M3
  Add missing license header to ReftableDatabase
  Add missing license header to UploadPackRefSortingForReachabilityTest
  [spotbugs] Fix potential NPE in FSTest
  Remove unused import in CreateFileSnapshotBenchmark
  RepositoryCache: don't require HEAD in git repositories
  FileRepository: cleanup refs outside refs/ on reftable conversion
  Update Orbit to S20191118194249 for 2019-12 M3
  Silence API error for new method in ReachabilityChecker
  UploadPackTest: Fix unused parameter in checkUnadvertisedIfUnallowed
  BitmappedReachabilityChecker: Use only one bitmap for the whole check
  checkNotAdvertisedWants: Be lazy converting Ref to RevCommit
  ReachabilityChecker: Receive a Stream instead of a Collection
  UploadPack: Prioritize references for non-advertised wants checks
  ReceivePack: Fix name hiding of 'atomic' member
  ReceivePack: Remove unnecessarily nested else-clause

+++ org.eclipse.jgit.test/tst/org/eclipse/jgit/util/FSTest.java
@@ -254,7 +254,9 @@ public class FSTest {
 						formatter.format(t1.toInstant()),
 						Long.valueOf(resolutionNs)), t2.compareTo(t1) > 0);
 			} finally {
-				Files.delete(f);
+				if (f != null) {
+					Files.delete(f);
+				}
 			}
 		}
 	}
